requirements.txt
requests
bs4
pony
bottle
ipdb
pyyaml
click
soupsieve
yq
